Redesigned and re-implemented the CSV file import in Service Layer
- Used the CSVLoader class to load the uploaded file data and convert it
to a 2D-list of strings (each row of the list representing an object as
a string). This list is then used in the CourseImporter and
StudentRegistrationImporter to create the corresponding list of objects
that will be saved in the database.
- Used the Strategy and Factory design pattern for an easier future
extension of the Fileloader (e.g. ExcelLoader). The idea behind this
approach is that every concrete FileLoader will convert the file
contents to a 2D-list of strings that will be used by the importers to
create the objects that will be saved in the database.