Implement the back-end for grade distribution chart
In a nutshell, the back-end calculates the grade distribution and
structures the results into a JSON string containing an array of json
objects that is later passed to the front-end.

Note: The data is passed from the back end to the front-end using a json string. This design decision aims to maintain a separation between data processing and visualization concerns.